requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
filelock==3.7.0
huggingface-hub==0.6.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
numpy==1.21.6
opencc-python-reimplemented==0.1.6
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
sentencepiece==0.1.96
six==1.16.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
